fix(smoke): harden cross-host evidence runner
